Revert "AE-2164: New energiatodistus signing UI"
