+++ etp-front/.dockerignore
@@ -9,13 +9,13 @@
 !.storybook/
 !__snapshots__/
 
-!babel.config.cjs
+!babel.config.js
 !fill-sv-localization.js
 !init-mocha.js
 !jsconfig.json
 !modheaders.json
-!postcss.config.cjs
-!svelte.config.cjs
+!postcss.config.js
+!svelte.config.js
 !tailwind.config.js
 !webpack.config.js
