+++ etp-front/jest_setup.js
@@ -1,5 +1,1 @@
 import '@testing-library/jest-dom';
-import 'cross-fetch/polyfill';
-import { configure } from '@testing-library/svelte';
-
-configure({ testIdAttribute: 'data-cy' });
