+++ etp-front/package.json
@@ -17,7 +17,6 @@
     "build:docs": "jsdoc -r -c ./jsdoc/conf.json",
     "watch:docs": "nodemon --exec npm run build:docs --watch src"
   },
-  "type": "module",
   "browserslist": [
     "defaults"
   ],
@@ -45,7 +44,6 @@
     "concurrently": "^8.2.2",
     "copy-webpack-plugin": "^12.0.0",
     "cross-env": "^7.0.2",
-    "cross-fetch": "^4.0.0",
     "css-loader": "^7.0.0",
     "cssnano": "^6.0.0",
     "generate-json-webpack-plugin": "^2.0.0",
@@ -53,7 +51,6 @@
     "http-server": "^14.1.1",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "jest-fetch-mock": "^3.0.3",
     "jest-image-snapshot": "^6.4.0",
     "jest-junit": "^16.0.0",
     "mini-css-extract-plugin": "^2.8.1",
@@ -90,6 +87,7 @@
     "monet": "^0.9.1",
     "qs": "^6.9.7",
     "quill": "^1.3.7",
+    "quill-image-drop-module": "^1.0.3",
     "quill-magic-url": "^4.0.0",
     "ramda": "^0.29.0",
     "svelte-i18n": "^4.0.0",
