--- etp-front/src/components/text-editor/text-editor.test.js
@@ -1,25 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-import { beforeAll, expect, test } from '@jest/globals';
-import { render, screen } from '@testing-library/svelte';
-
-import TextEditor from './text-editor.svelte';
-import { setupI18n } from '@Language/i18n.js';
-
-beforeAll(() => {
-  setupI18n();
-});
-
-test('TextEditor renders correctly', () => {
-  render(TextEditor, {
-    id: 'test',
-    name: 'test',
-    label: 'test',
-    required: true,
-    model: 'test content'
-  });
-
-  const content = screen.getByText('test content');
-  expect(content).toBeInTheDocument();
-});
