--- etp-front/src/pages/energiatodistus/signing/SigningDialog.stories.js
@@ -1,22 +1,0 @@
-import SigningDialog from './SigningDialog.svelte';
-import { energiatodistus2018 } from '@Pages/energiatodistus/empty';
-
-export default { title: 'SigningDialog' };
-
-export const signingWithCardNoConnection = () => ({
-  Component: SigningDialog,
-  props: {
-    energiatodistus: energiatodistus2018(),
-    selection: 'card',
-    allowSelection: true
-  }
-});
-
-export const signingWithoutCard = () => ({
-  Component: SigningDialog,
-  props: {
-    energiatodistus: energiatodistus2018(),
-    selection: 'system',
-    allowSelection: true
-  }
-});
