--- etp-front/src/pages/energiatodistus/signing/SigningInstructions.svelte
@@ -1,3 +1,0 @@
-<div>
-  <p>Allekirjoittamiseen liittyvät ohjeet tulevat tähän</p>
-</div>
