+++ etp-front/webpack.config.js
@@ -1,22 +1,19 @@
-import { format } from 'date-fns/format';
+const { format } = require('date-fns/format');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const BundleAnalyzerPlugin =
+  require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
-import { CleanWebpackPlugin } from 'clean-webpack-plugin';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import GenerateJsonPlugin from 'generate-json-webpack-plugin';
-import CopyWebpackPlugin from 'copy-webpack-plugin';
-import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+const GenerateJsonPlugin = require('generate-json-webpack-plugin');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
 
-import path, { dirname } from 'path';
-import { fileURLToPath } from 'url';
+const path = require('path');
 
 const mode = process.env.NODE_ENV || 'development';
 const prod = mode === 'production';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-export default {
+module.exports = {
   entry: {
     bundle: ['./src/main.js']
   },
@@ -66,18 +63,12 @@ export default {
     rules: [
       {
         test: /\.m?js$/,
-        resolve: {
-          fullySpecified: false
-        },
         exclude: /node_modules/,
         include: [path.resolve(__dirname, 'src'), /svelte/],
         use: ['babel-loader']
       },
       {
         test: /\.svelte$/,
-        resolve: {
-          fullySpecified: false
-        },
         include: [path.resolve(__dirname, 'src'), /svelte/],
         use: [
           'babel-loader',
